requirements.txt
tensorflow-macos==2.13.0
tensorflow-metal==1.0.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
wandb==0.15.8
jupyter==1.0.0
ipykernel==6.25.1
tqdm==4.66.1
audioread==3.0.0
